requirements.txt
absl-py==2.1.0
chex==0.1.85
dm-haiku==0.0.11
jax==0.4.23
numpy==1.26.3
